part 4 d

+++ part 4/controllers/blogs.js
@@ -1,14 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs.map(blog => blog.toJSON()))
 })
 
 blogsRouter.get('/:id', async (request, response) => {
-  const blog = await Blog.findById(request.params.id)
+  const blog = await Blog.findById(request.params.id).populate('user', {
+    username: 1,
+    name: 1,
+  })
   if (blog) {
     response.json(blog.toJSON())
   } else {
@@ -17,24 +19,15 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const body = request.body
+  if (!request.token || !request.decodedToken) {
+    return response.status(401).json({ error: 'missing or invalid token' })
+  }
 
-  const user = await User.findById(body.userId)
+  const body = request.body
+  const user = request.user
 
   if (!body.likes) {
     body.likes = 0
-  }
-
-  if (!body.title) {
-    return response.status(400).json({
-      error: 'title missing',
-    })
-  }
-
-  if (!body.author) {
-    return response.status(400).json({
-      error: 'author missing',
-    })
   }
 
   const blog = new Blog({
@@ -48,16 +41,47 @@ blogsRouter.post('/', async (request, response) => {
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
-  response.json(savedBlog.toJSON())
+
+  // Populate user field on the returned blog
+  const populatedBlog = await savedBlog
+    .populate('user', { username: 1, name: 1 })
+    .execPopulate()
+
+  if (populatedBlog) {
+    response.status(201).json(savedBlog.toJSON())
+  } else {
+    response.status(400).end()
+  }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  console.log(Blog.findByIdAndRemove(request.params.id))
+  if (!request.token || !request.decodedToken) {
+    return response.status(401).json({ error: 'missing or invalid token' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(400).json({
+      error: 'blog already deleted',
+    })
+  }
+
+  if (blog.user.toString() !== request.user.id.toString()) {
+    return response.status(401).json({
+      error: 'blogs can be deleted only by the user who added the blog',
+    })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
+  if (!request.token || !request.decodedToken) {
+    return response.status(401).json({ error: 'missing or invalid token' })
+  }
+
   const body = request.body
 
   const blog = {
@@ -71,7 +95,12 @@ blogsRouter.put('/:id', async (request, response) => {
     new: true,
   })
 
-  response.json(updatedBlog.toJSON())
+  // Populate user field on the returned blog
+  const populatedBlog = await updatedBlog
+    .populate('user', { username: 1, name: 1 })
+    .execPopulate()
+
+  response.json(populatedBlog.toJSON())
 })
 
 module.exports = blogsRouter
